+++ plugins/actix-web/src/lib.rs
@@ -7,7 +7,8 @@ pub mod web;
 
 pub use self::web::{Resource, Route, Scope};
 pub use paperclip_macros::{
-    api_v2_errors, api_v2_operation, delete, get, post, put, Apiv2Schema, Apiv2Security,
+    api_v2_errors, api_v2_errors_overlay, api_v2_operation, delete, get, post, put, Apiv2Schema,
+    Apiv2Security,
 };
 
 use self::web::{RouteWrapper, ServiceConfig};